Change how files are stored
We now use a separate CF for blobs. Files are stored uncompressed and we
use the DeflateCompressor on the column family. Cassandra's compression
seems to be faster than the Python client I'm using (even with the large
buffer allocations). If this turns out to be a bad idea, we still
technically support storage of compressed blobs. And, the compressor on
the column family is easy enough to change.